Add filters. (#3)
Instead of strings we now have a `Filter` class and many top-level
convenience methods to create them.

With this patch it is now possible to use more operators and also to
create logical combination ('or', 'and') which was previously not
possible. Similarly, it was not possible to test for 'null' values.